requirements.txt
# 项目依赖

# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库依赖
redis==5.0.1
sqlalchemy==2.0.23

# LLM依赖
requests==2.33.1

# 安全工具集成依赖
python-nmap==0.6.1
pyowaspzap==0.0.1
python-metasploit3==1.0.1

# 工具执行依赖
subprocess32==3.5.4
paramiko==3.4.0

# 报告生成依赖
jinja2==3.1.2
markdown==3.5.0

# 监控依赖
prometheus-client==0.19.0

# 测试依赖
pytest==7.4.3
httpx==0.25.2

# 开发依赖
black==23.11.0
flake8==6.1.0
isort==5.12.0